requirements.txt
aiogram==3.13.1
httpx==0.27.2
aiosqlite==0.20.0
cryptography==43.0.1
python-dotenv==1.0.1
